backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
google-generativeai==0.8.5
python-dotenv==1.0.0
pydantic==1.10.12 